+++ js/admin.js
@@ -1,7 +1,7 @@
 $(function(){
-    /* toogle admin form sidebar */
-    $('#kUtRL h5').toggleWithLegend(
-        $('#kUtRL').children().not('h5'),
-        {cookie:'dcx_kUtRL_admin_form_sidebar',legend_click:true}
-    );
+  $('#form-entries .checkboxes-helpers').each(function () {
+    dotclear.checkboxesHelpers(this, undefined, '#form-entries td input[type=checkbox]', '#form-entries #do-action');
+  });
+  $('#form-entries td input[type=checkbox]').enableShiftClick();
+  dotclear.condSubmit('#form-entries td input[type=checkbox]', '#form-entries #do-action');
 });